pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polaroids"
version = "0.1.0"
description = "Build pipeline for the Polaroids for Mom Mother's Day screensaver"
requires-python = ">=3.13"
dependencies = [
  "typer>=0.12,<0.13",
  # typer 0.12.x is only compatible with click < 8.2 (typer release notes
  # 0.15.4 added the < 8.2 pin retroactively; we pin it explicitly here so
  # a fresh install on a machine where pip's default resolution picks
  # click 8.2+ does not yield the make_metavar(ctx) TypeError.
  "click>=8.0,<8.2",
  "Pillow==12.2.0",
  "pillow-heif==1.3.0",
  "piexif==1.1.3",
  "openai==2.36.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
]

[project.scripts]
polaroids = "polaroids.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["polaroids*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-x -ra"
